APIClientでトークンを保持する
