+++ TurmericTests/TurmericTests.swift
@@ -56,9 +56,7 @@ class TurmericTests: XCTestCase {
     func testUserLogin() {
         waitUntil { done in
             User.authenticate(parameters: ["user" : ["email" : "syuta_ogido@yahoo.co.jp", "password" : "testtest"]]) { response in
-                XCTAssertEqual("testUser", CurrentUser.name)
-                XCTAssertEqual("test@test.com", CurrentUser.email)
-                XCTAssertEqual("ThisIsAuthToken", CurrentUser.loadToken())
+                XCTAssertEqual("ThisIsAuthToken", APIClient.loadToken())
                 done()
             }
         }
